Add BH1750 and S12SD drivers
Added some basic drivers for the BH1750 and S12SD peripherals
used in this project. The S12SD is a UV sensor that communicates over an
analog line. The BH1750 is a light sensor that communicates over I2C.
Both drivers are highly coupled to the Pico C SDK but that may change in
the future.

I also implemented a simple main function that polls the devices to
ensure they work.